// packages/rapidroute/src/types/wiki/ImageSearch.ts
// Generated by https://quicktype.io

export interface WikiImage {
  batchcomplete: string
  query: Query
}

export interface Query {
  pages: Pages
}

export type Pages = Record<string, Missing | Found>

export interface Missing {
  ns: number
  title: string
  missing: string
  imagerepository: string
}

export interface Found {
  pageid: number
  ns: number
  title: string
  imagerepository: string
  imageinfo: Imageinfo[]
}

export interface Imageinfo {
  thumburl: string
  thumbwidth: number
  thumbheight: number
  responsiveUrls: ResponsiveUrls
  url: string
  descriptionurl: string
  descriptionshorturl: string
}

export interface ResponsiveUrls {
  "1.5": string
  "2": string
}